kafka_streaming/requirements.txt
six
kafka-python>=2.0.2
pyspark==3.5.7
findspark==2.0.1

